feat(perps): add PerpsGuideEntryPopup for enhanced user guidance on market detail navigation
