+++ apps/mobile/src/screens/Perps/components/PerpsGuideEntryPopup.tsx
